requirements.txt
black==23.3.0
colorama==0.4.6
future==0.18.3
mando==0.6.4
mypy==1.2.0
pylint==2.17.2
PyYAML==5.3.1
types-PyYAML==6.0.12.9
typing_extensions==4.5.0
watchdog==3.0.0